Refactor of syncSite
