+++ docker/CMSRucioClient/scripts/syncSite.py
@@ -5,38 +5,20 @@ Command line tool for registering a CMS dataset into rucio
 
 from __future__ import absolute_import, division, print_function
 
-import json
 import multiprocessing
 import re
 from argparse import ArgumentParser
-from subprocess import Popen, PIPE
 
-from gfal2 import Gfal2Context, GError
 import rucio.rse.rsemanager as rsemgr
-from rucio.client.didclient import DIDClient
-from rucio.client.replicaclient import ReplicaClient
-from rucio.common.exception import DataIdentifierAlreadyExists
-from rucio.common.exception import RucioException
-from rucio.common.exception import FileAlreadyExists
+from CMSRucio import CMSRucio, DEFAULT_DASGOCLIENT, das_go_client
+from gfal2 import GError, Gfal2Context
 
 BLOCKREPLICAS_URL = "https://cmsweb.cern.ch/phedex/datasvc/json/prod/blockreplicas"
 DEBUG_FLAG = False
 DEFAULT_SCOPE = 'cms'
-DEFAULT_DASGOCLIENT = '/usr/bin/dasgoclient'
 DEFAULT_LIMIT = 10
 
-def das_go_client(query, dasgoclient=DEFAULT_DASGOCLIENT):
-    """
-    just wrapping the dasgoclient command line
-    """
-    proc = Popen([dasgoclient, '-query=%s' % query, '-json'], stdout=PIPE)
-    output = proc.communicate()[0]
-    if DEBUG_FLAG:
-        print('DEBUG:' + output)
-    return json.loads(output)
-
-
-class DatasetSync(object):
+class DatasetSync(CMSRucio):
     """
     Synchronize the replica of a given dataset at a PhEDEx site to
     the corresponding Rucio site.
@@ -48,15 +30,16 @@ class DatasetSync(object):
            :param dataset: Name of the PhEDEx dataset to synchronize with Rucio.
            :param pnn: PhEDEx node name to filter on for replica information.
         """
+
+        super(DatasetSync, self).__init__(account=None, auth_type=None, scope=scope, dry_run=dry_run)
+
         self.phedex_dataset = dataset
-        self.pnn = pnn 
+        self.pnn = pnn
         if rse is None:
-            rse = pnn 
+            rse = pnn
         self.rse = rse
-        self.scope = scope
         self.check = check
         self.lifetime = lifetime
-        self.dry_run = dry_run
         self.dasgoclient=dasgoclient
 
         self.rucio_datasets = {}
@@ -64,8 +47,6 @@ class DatasetSync(object):
         self.url = ''
 
         self.get_phedex_metadata()
-        self.didc = DIDClient()
-        self.repc = ReplicaClient()
         self.get_rucio_metadata()
         self.get_global_url()
 
@@ -125,7 +106,7 @@ class DatasetSync(object):
         """
         print("Initializing Rucio... getting the list of blocks and files at %s"
               % self.rse)
-        replica_info = self.repc.list_replicas([{"scope": self.scope,
+        replica_info = self.rc.list_replicas([{"scope": self.scope,
                                                  "name": self.phedex_dataset}],
                                                rse_expression="rse=%s" % self.rse)
         replica_files = set()
@@ -150,17 +131,17 @@ class DatasetSync(object):
         Create the container, the datasets and attach them to the container.
         """
         print("Registering...")
-        self.register_container()
+        self.register_container(dataset=self.dataset, lifetime=self.lifetime)
         # First, iterate through all the known PhEDEx blocks.
         for block_name, block_info in self.blocks.items():
             if block_name not in self.rucio_datasets:
-                self.register_dataset(block_name)
+                self.register_dataset(block_name, dataset=self.dataset, lifetime=self.lifetime)
             rucio_dataset_info = self.rucio_datasets.setdefault(block_name, {})
             files_to_attach = []
             for filename, filemd in block_info.items():
                 if filename not in rucio_dataset_info:
                     files_to_attach.append(filemd)
-            self.register_replicas(files_to_attach)
+            self.register_replicas(rse=self.rse, replicas=files_to_attach)
             self.attach_files([i['name'] for i in files_to_attach], block_name)
         # We do not detach files, but perhaps remove the replicas.
         for dataset_name, dataset_info in self.rucio_datasets.items():
@@ -169,112 +150,8 @@ class DatasetSync(object):
             for filename, filemd in dataset_info.items():
                 if filename not in block_info:
                     replicas_to_delete.append(filemd)
-            self.delete_replicas(replicas_to_delete)
+            self.delete_replicas(rse=self.rse, replicas=replicas_to_delete)
         print("All datasets, blocks and files registered")
-
-    def register_container(self):
-        """
-        Create the container.
-        """
-
-        print("registering container %s:%s" % (self.scope, self.phedex_dataset))
-        if self.dry_run:
-            print(' Dry run only. Not creating container.')
-            return
-
-        try:
-            self.didc.add_container(scope=self.scope, name=self.phedex_dataset, lifetime=self.lifetime)
-        except DataIdentifierAlreadyExists:
-            print("Container %s already exists" % self.phedex_dataset)
-
-    def register_dataset(self, block):
-        """
-        Create the dataset and attach them to the container
-        """
-        print("registering dataset %s" % block)
-
-        if self.dry_run:
-            print(' Dry run only. Not creating dataset.')
-            return
-
-        try:
-            self.didc.add_dataset(scope=self.scope, name=block, lifetime=self.lifetime)
-        except DataIdentifierAlreadyExists:
-            print(" Dataset %s already exists" % block)
-
-        try:
-            print("attaching dataset %s to container %s" % (block, self.phedex_dataset))
-            self.didc.attach_dids(scope=self.scope, name=self.phedex_dataset,
-                                  dids=[{'scope': self.scope, 'name': block}])
-        except RucioException:
-            print(" Dataset already attached")
-
-    def attach_files(self, lfns, block):
-        """
-        Attach the file to the container
-        """
-        if not lfns:
-            return
-
-        if self.dry_run:
-            print(' Dry run only. Not attaching files.')
-            return
-
-        try:
-            print("attaching files: %s" % ", ".join(lfns))
-            self.didc.attach_dids(scope=self.scope, name=block,
-                                  dids=[{'scope': self.scope, 'name': lfn} for lfn in lfns])
-        except FileAlreadyExists:
-            print("File already attached")
-
-    def register_replicas(self, replicas):
-        """
-        Register file replica.
-        """
-        if not replicas:
-            return
-
-        print("registering files in Rucio: %s" % ", ".join([filemd['name'] for filemd in replicas]))
-
-        if self.dry_run:
-            print(' Dry run only. Not registering files.')
-            return
-
-        if self.check:
-            filtered_replicas = []
-            for filemd in replicas:
-                if self.check_storage(filemd):
-                    filtered_replicas.append(filemd)
-            replicas = filtered_replicas
-
-        self.repc.add_replicas(rse=self.rse, files=[{
-                'scope': self.scope,
-                'name': filemd['name'],
-                'adler32': filemd['checksum'],
-                'bytes': filemd['size'],
-               } for filemd in replicas])
-
-    def delete_replicas(self, replicas):
-        """
-        Delete replicas from the current RSE.
-        """
-        if not replicas:
-            return
-
-        print("Deleting files from %s in Rucio: %s" % (self.rse,
-              ", ".join([filemd['name'] for filemd in replicas])))
-
-        if self.dry_run:
-            print(" Dry run only.  Not deleting replicas.")
-            return
-
-        try:
-            self.repc.delete_replicas(rse=self.rse, files=[{
-               'scope': self.scope,
-               'name': filemd['name'],
-             } for filemd in replicas])
-        except rucio.common.exception.AccessDenied:
-            print("Permission denied in deleting replicas: %s" % ", ".join([filemd['name'] for filemd in replicas]))
 
     def check_storage(self, filemd):
         """
